initial restructure, updated readme to reflect new installation procedure
