+++ normflow/geometry.py
